% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_microsoftSQL.R
\name{microsoftSQL_setup_server}
\alias{microsoftSQL_setup_server}
\title{microsoftSQL Setup Server}
\usage{
microsoftSQL_setup_server(id)
}
\arguments{
\item{id}{The Module namespace}
}
\value{
microsoftSQL connection variables
}
\description{
microsoftSQL Setup Server
}
\keyword{internal}
