% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_microsoftSQL.R
\name{microsoftSQL_setup_ui}
\alias{microsoftSQL_setup_ui}
\title{microsoftSQL Setup UI}
\usage{
microsoftSQL_setup_ui(id)
}
\arguments{
\item{id}{The module namespace}
}
\value{
The microsoftSQL Setup UI
}
\description{
This module is designed to guide a user through the process of authenticating your database
}
\keyword{internal}
